user-grpc/requirements.txt
grpcio
grpcio-tools
pymongo==3.5.1
sentry-sdk==0.16.2
fluent-logger
pinject